Port r730523 from tcnative trunk to tc trunk. Add Thread.java and Buffer.java. Change SSL.randSet() javadoc to the tcnative one. Add Socket.sendib() and Socket.sendibb from tcnative.
All renames methods and method signature changes
are compatible, because those methods are not used
by Tomcat.

Changes of return types are compatible, because
the return values have not been used.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@730544 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/tomcat/jni/Stdlib.java
@@ -67,7 +67,7 @@ public class Stdlib {
      * @param num Number of elements.
      * @param sz Length in bytes of each element.
      */
-    public static native long calloc(long num, int sz);
+    public static native long calloc(int num, int sz);
 
     /**
      * Deallocates or frees a memory block.
